requirements.txt
reportlab==4.2.2
vnstock==0.2.9.2.2
vnstock3==0.3.0.7
mplfinance==0.12.10b0
openai==1.50.2
PyMuPDF==1.24.10
python-dotenv==1.0.0

flask==2.2.3
pandas==2.0.1
numpy==1.24.3

pandas==2.0.1
pandas-datareader==0.10.0
pandas-gbq==0.17.9
pytz==2022.7.1
requests==2.30.0
matplotlib==3.6.2
google-cloud-bigquery==3.5.0
Werkzeug==2.2.2